refactor(runtime): move subagent bridge into package as nanobot.runtime.bridge (#599)
Moves scripts/eeepc_self_evolving_subagent_bridge.py (1,484 LOC, canonical
implementation) into nanobot/runtime/bridge.py; the script path becomes a
10-line wrapper that calls nanobot.runtime.bridge.cli_main() so the
systemd unit and deploy_release.sh's file-copy step stay unchanged
(retiring that mechanism is #601). Deletes the stale, drifted
host/eeepc/libexec/eeepc-self-evolving-subagent-bridge.py stub (314 LOC,
1,224 diff lines from canonical) and updates install.sh to install the
wrapper from scripts/ so a fresh install and a later deploy always agree.

The three structural bugs documented in blueprint/bridge-feedback-loop-fix/
(commits_pushed misreport, previous-attempts matching, missing
backlog_title anchor) were already fixed in the code prior to this move;
folds that content into docs/changes/599-bridge-into-package/proposal.md
and deletes the blueprint/ folder. Existing regression tests for those
fixes (test_commits_pushed.py, test_lessons_feedback_loop.py, etc.) are
repointed at the new module path; adds test_bridge_wrapper.py for the
wrapper itself.
