+++ docs/specs/README.md
@@ -49,7 +49,7 @@ where its sources went.
 | Capability | Source docs folded → `.legacy/` | Kept as reference / runbook |
 |---|---|---|
 | `self-evolving-runtime` | runtime-subagent directive (from AGENTS.md) | EEEBOT_SELF_IMPROVING_RUNTIME_OPERATING_CONTRACT, EEEBOT_INSIGHT_HYPOTHESIS_LOOP_CLOSURE, EEEBOT_BUDGET_AND_REWARD_MODEL, EEEBOT_EXPERIMENT_AND_OUTCOME_CONTRACT (detail behind the spec) |
-| `subagent-bridge` | — | SYSTEM_OPERATION_REFERENCE §6, `scripts/eeepc_self_evolving_subagent_bridge.py` (authoritative) |
+| `subagent-bridge` | — | SYSTEM_OPERATION_REFERENCE §6, `nanobot/runtime/bridge.py` (authoritative) |
 | `promotion-and-release` | SOURCE_OF_TRUTH_AND_PROMOTION_POLICY, PROMOTION_GATE_SPEC, RELEASE_ARTIFACT_AND_ROLLBACK_CONTRACT, BRANCH_AND_RELEASE_CHANNEL_POLICY, CHANGE_PROPAGATION_MODEL, VERSION_AND_PROVENANCE_MODEL, DRIFT_BUDGET_AND_RECONCILIATION_POLICY, VALIDATION_HOOKS_PLAN, DEPLOY_DECISION_MATRIX | — |
 | `host-runtime` | EEEPC_AGENT_RUNTIME_INSTRUCTIONS, EEEPC_RUNTIME_STATE_AUTHORITY_USAGE, HOST_CAPABILITY_POLICY, HOST_GITHUB_SYNC_ARCHITECTURE, HOST_WORKSPACE_ARTIFACT_TRIAGE, SAFE_BOOTSTRAP_FROM_SCRATCH, BASE_CONFIGURATION_PROFILE | EEEPC_DEPLOY_VERIFY_ROLLBACK_RUNBOOK, EEEPC_APPLY_OK_OPERATOR_RUNBOOK (runbooks) |
 | `model-routing` | MODEL_ROUTING_FALLBACK_V1 | — |
