+++ host/eeepc/scripts/install.sh
@@ -157,6 +157,15 @@ install_libexec() {
     run chmod +x "/usr/local/libexec/$name"
     log "  ✓ $name"
   done
+
+  # Subagent bridge: no libexec stub exists anymore (#599) — the canonical
+  # implementation lives in nanobot/runtime/bridge.py and this thin wrapper
+  # at scripts/eeepc_self_evolving_subagent_bridge.py is what deploy_release.sh
+  # also installs, so a fresh install and a later deploy always agree.
+  run cp "$REPO_ROOT/scripts/eeepc_self_evolving_subagent_bridge.py" \
+         /usr/local/libexec/eeepc-self-evolving-subagent-bridge.py
+  run chmod +x /usr/local/libexec/eeepc-self-evolving-subagent-bridge.py
+  log "  ✓ eeepc-self-evolving-subagent-bridge.py (from scripts/, wrapper)"
 }
 
 # ---------------------------------------------------------------------------
