+++ tests/test_mutation_surfaces.py
@@ -11,7 +11,7 @@ from pathlib import Path
 
 import pytest
 
-_BRIDGE_PATH = Path(__file__).parent.parent / 'scripts' / 'eeepc_self_evolving_subagent_bridge.py'
+_BRIDGE_PATH = Path(__file__).parent.parent / 'nanobot' / 'runtime' / 'bridge.py'
 
 
 def _extract_fn(name: str, extra_setup: str = '') -> object:
