+++ tests/test_structured_lessons.py
@@ -18,7 +18,7 @@ sys.path.insert(0, str(Path(__file__).parent.parent))
 
 def _load_bridge_ns(*names: str) -> dict:
     """Extract named functions from bridge without triggering imports."""
-    bridge_path = Path(__file__).parent.parent / "scripts" / "eeepc_self_evolving_subagent_bridge.py"
+    bridge_path = Path(__file__).parent.parent / "nanobot" / "runtime" / "bridge.py"
     source = bridge_path.read_text()
     ns: dict = {"re": re, "Path": Path, "json": json, "datetime": datetime}
     tree = ast.parse(source)
